# Conflicts:
#	example.json

+++ src/main/java/kz/syllabus/repository/DisciplineRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DisciplineRepository extends JpaRepository<Discipline, Long> {
+public interface DisciplineRepository extends JpaRepository<Discipline, Integer> {
 
 }
